+++ src/command.ts
@@ -6,17 +6,6 @@ import { client } from './ecs/client'
 import { Variables } from './types/configuration'
 
 export class AwsCommand extends Command {
-  // Ensure AWS_PROFILE is set, otherwise we may connect to the wrong account
-  // TODO: Figure out how to run this by default
-  ensure_aws_profile() {
-    const AWS_PROFILE = process.env.AWS_PROFILE || undefined
-    if (AWS_PROFILE === '' || AWS_PROFILE === undefined) {
-      this.error('Please ensure AWS_PROFILE environment variable is set.')
-    }
-
-    return AWS_PROFILE
-  }
-
   configWithVariables(variables: Variables = {}) {
     const initialVariables = {
       ...this.variables(),
@@ -27,8 +16,6 @@ export class AwsCommand extends Command {
   }
 
   ecs_client() {
-    this.ensure_aws_profile()
-
     return client
   }